requirements.txt
agno>=1.4.0
openai>=1.30.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pytest>=8.0.0
